Refactor shared functions into utility classes
